Update GridView.swift
"Length" appears to be deprecated or maybe is defined in a file not included in this repo? 

It is initialized with an Int in GridView.init(), so I tried typealiasing it to Int first.  The compiler complains that it expects a CGFloat in several places though, so I left it as that.